@@ -14,15 +14,29 @@ description: A new Flutter project.
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.2.2 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-
   intl: ^0.15.8
   url_launcher: ^5.1.2
   qr: ^1.2.0
+  uuid: 2.0.1
+  shared_preferences: ^0.5.3+4
+  flutter_secure_storage: ^3.2.1+1
+  provider: ^3.1.0
+  rxdart: ^0.22.2
+  yaml: ^2.1.16
+  barcode_scan: any
+  http: ^0.12.0+2
+  sqflite: ^1.1.6
+  path_provider: ^1.3.0
+  mobx: ^0.3.7
+  flutter_mobx: 0.3.0+1
+  flutter_slidable: ^0.5.3
+  share: ^0.6.2+1
+  esys_flutter_share: ^1.0.2
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
@@ -31,6 +45,8 @@ dependencies:
 dev_dependencies:
   flutter_test:
     sdk: flutter
+  build_runner: ^1.3.1
+  mobx_codegen: 0.3.3+1
 
 
 # For information on the generic Dart part of this file, see the
@@ -46,6 +62,7 @@ flutter:
 
   assets:
     - assets/images/
+    - assets/node_list.yml
     - assets/text/
 
   fonts:
