@@ -1,12 +1,44 @@
-const String welcomeRoute = '/';
-const String newWalletRoute = '/new_wallet';
-const String setupPinRoute = '/setup_pin_code';
-const String newWalletFromWelcomeRoute = '/new_wallet_from_welcome';
-const String restoreFromWelcomeRoute = '/restore_from_welcome';
-const String restoreRoute = '/restore';
-const String restoreSeedKeysRoute = '/restore_seed_keys';
-const String seedRoute = '/seed';
-const String restoreFromSeedRoute = '/restore_from_seed';
-const String restoreFromKeysRoute = '/restore_from_keys';
-const String disclaimerRoute = '/disclaimer';
-const String seedAlertRoute = '/seed_alert';
+class Routes {
+  static const welcome = '/welcome';
+  static const newWallet = '/new_wallet';
+  static const setupPin = '/setup_pin_code';
+  static const newWalletFromWelcome = '/new_wallet_from_welcome';
+  static const restoreFromWelcome = '/restore_from_welcome';
+  static const seed = '/seed';
+  static const restoreOptions = '/restore_options';
+  static const restoreOptionsFromWelcome = '/restore_options_from_welcome';
+  static const restoreWalletOptions = '/restore_seed_keys';
+  static const restoreWalletOptionsFromWelcome = '/restore_wallet_options_from_welcome';
+  static const restoreWalletFromSeed = '/restore_wallet_from_seed';
+  static const restoreWalletFromKeys = '/restore_wallet_from_keys';
+  static const dashboard = '/dashboard';
+  static const send = '/send';
+  static const transactionDetails = '/transaction_info';
+  static const receive = '/receive';
+  static const newSubaddress = '/new_subaddress';
+  static const disclaimer = '/disclaimer';
+  static const readDisclaimer = '/read_disclaimer';
+  static const seedAlert = '/seed_alert';
+  static const walletList = '/wallet_list';
+  static const auth = '/auth';
+  static const nodeList = '/node_list';
+  static const newNode = '/new_node_list';
+  static const home = '/home';
+  static const login = '/login';
+  static const splash = '/splash';
+  static const accountList = '/account_list';
+  static const accountCreation = '/account_new';
+  static const addressBook = '/address_book';
+  static const pickerAddressBook = '/picker_address_book';
+  static const addressBookAddContact = '/address_book_add_contact';
+  static const showKeys = '/show_keys';
+  static const exchangeConfirm = '/exchange_confirm';
+  static const tradeHistory = '/trade_history';
+  static const tradeDetails = '/trade_details';
+  static const exchangeFunds = '/exchange_funds';
+  static const exchangeTrade = '/exchange_trade';
+  static const subaddressList = '/subaddress_list';
+  static const restoreWalletFromSeedDetails = '/restore_from_seed_details';
+  static const exchange = '/exchange';
+  static const settings = '/settings';
+}
