@@ -1,24 +1,44 @@
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/cupertino.dart';
+import 'package:provider/provider.dart';
+import 'package:cake_wallet/src/stores/user/user_store.dart';
 import 'package:cake_wallet/src/screens/pin_code/pin_code.dart';
+import 'package:cake_wallet/src/screens/base_page.dart';
 
-class SetupPinCode extends StatefulWidget {
-  final Function(BuildContext) onPinCodeSetup;
-  
-  SetupPinCode(this.onPinCodeSetup);
+class SetupPinCodePage extends BasePage {
+  final Function(BuildContext, String) onPinCodeSetup;
 
   @override
-  _SetupPinCodeState createState() => _SetupPinCodeState();
+  String get title => 'Setup PIN';
+
+  SetupPinCodePage({this.onPinCodeSetup});
+
+  @override
+  Widget body(BuildContext context) =>
+      SetupPinCodeForm(onPinCodeSetup: onPinCodeSetup, hasLengthSwitcher: true);
 }
 
-class _SetupPinCodeState<WidgetType extends SetupPinCode> extends PinCodeState<WidgetType> {
+class SetupPinCodeForm extends PinCodeWidget {
+  final Function(BuildContext, String) onPinCodeSetup;
+  final bool hasLengthSwitcher;
+
+  SetupPinCodeForm(
+      {@required this.onPinCodeSetup, @required this.hasLengthSwitcher});
+
+  @override
+  _SetupPinCodeFormState createState() => _SetupPinCodeFormState();
+}
+
+class _SetupPinCodeFormState<WidgetType extends SetupPinCodeForm>
+    extends PinCodeState<WidgetType> {
+
   bool isEnteredOriginalPin() => !(_originalPin.length == 0);
   Function(BuildContext) onPinCodeSetup;
   List<int> _originalPin = [];
-  static final backArrowImage = Image.asset('assets/images/back_arrow.png');
-  
-  _SetupPinCodeState() {
+  UserStore _userStore;
+
+  _SetupPinCodeFormState() {
     title = "Enter your pin";
   }
 
@@ -29,44 +49,44 @@ class _SetupPinCodeState<WidgetType extends SetupPinCode> extends PinCodeState<W
       state.title = 'Enter your pin again';
       state.clear();
     } else {
-
       if (listEquals<int>(state.pin, _originalPin)) {
+        final String pin = state.pin.fold("", (ac, val) => ac + '$val');
+        _userStore.set(password: pin);
+
         showDialog(
-          context: context,
-          barrierDismissible: false,
-          builder: (BuildContext context) {
-            return AlertDialog(
-              content: Text("Your PIN has been set up successfully!"),
-              actions: <Widget>[
-                FlatButton(
-                  child: Text("OK"),
-                  onPressed: () {
-                    Navigator.of(context).pop();
-                    widget.onPinCodeSetup(context);
-                    reset();
-                  },
-                ),
-              ],
-            );
-          }
-        );
+            context: context,
+            barrierDismissible: false,
+            builder: (BuildContext context) {
+              return AlertDialog(
+                content: Text("Your PIN has been set up successfully!"),
+                actions: <Widget>[
+                  FlatButton(
+                    child: Text("OK"),
+                    onPressed: () {
+                      Navigator.of(context).pop();
+                      widget.onPinCodeSetup(context, pin);
+                      reset();
+                    },
+                  ),
+                ],
+              );
+            });
       } else {
         showDialog(
-          context: context,
-          builder: (BuildContext context) {
-            return AlertDialog(
-              content: Text("PIN is incorrect"),
-              actions: <Widget>[
-                FlatButton(
-                  child: Text("OK"),
-                  onPressed: () {
-                    Navigator.of(context).pop();
-                  },
-                ),
-              ],
-            );
-          }
-        );
+            context: context,
+            builder: (BuildContext context) {
+              return AlertDialog(
+                content: Text("PIN is incorrect"),
+                actions: <Widget>[
+                  FlatButton(
+                    child: Text("OK"),
+                    onPressed: () {
+                      Navigator.of(context).pop();
+                    },
+                  ),
+                ],
+              );
+            });
 
         reset();
       }
@@ -81,22 +101,8 @@ class _SetupPinCodeState<WidgetType extends SetupPinCode> extends PinCodeState<W
 
   @override
   Widget build(BuildContext context) {
-    return Scaffold(
-      backgroundColor: Colors.white,
-      appBar: CupertinoNavigationBar(
-        leading: ButtonTheme(
-          minWidth: double.minPositive,
-          child: FlatButton(
-            onPressed: (){Navigator.pop(context);},
-            child: backArrowImage
-          ),
-        ),
-        middle: Text('Setup Pin', style: TextStyle(fontSize: 16.0, fontWeight: FontWeight.w600),),
-        backgroundColor: Colors.white,
-        border: null,
-      ),
-      body: body(context)
-    );
-  }
+    _userStore = Provider.of<UserStore>(context);
 
+    return body(context);
+  }
 }
